+++ sumologic/resource_sumologic_monitors_library_monitor_test.go
@@ -249,6 +249,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 	testType := "MonitorsLibraryMonitor"
 	testContentType := "Monitor"
 	testMonitorType := "Logs"
+	testPlaybook := "This is a test playbook"
 	testIsDisabled := false
 	testQueries := []MonitorQuery{
 		{
@@ -306,6 +307,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 	testUpdatedType := "MonitorsLibraryMonitor"
 	testUpdatedContentType := "Monitor"
 	testUpdatedMonitorType := "Logs"
+	testUpdatedPlaybook := "This is an updated test playbook"
 	testUpdatedIsDisabled := true
 	testUpdatedQueries := []MonitorQuery{
 		{
@@ -377,6 +379,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.trigger_type", testTriggers[0].TriggerType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.time_range", testTriggers[0].TimeRange),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "notifications.0.notification.0.connection_type", testNotifications[0].Notification.(EmailNotification).ConnectionType),
+					resource.TestCheckResourceAttr("sumologic_monitor.test", "playbook", testPlaybook),
 				),
 			},
 			{
@@ -392,6 +395,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.trigger_type", testUpdatedTriggers[0].TriggerType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.time_range", testUpdatedTriggers[0].TimeRange),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "notifications.0.notification.0.connection_type", testUpdatedNotifications[0].Notification.(EmailNotification).ConnectionType),
+					resource.TestCheckResourceAttr("sumologic_monitor.test", "playbook", testUpdatedPlaybook),
 				),
 			},
 		},
@@ -582,6 +586,7 @@ resource "sumologic_monitor" "test" {
 		  }
 		run_for_trigger_types = [%s]
 	  }
+	playbook = "This is a test playbook"
 }`, testName, monitorType, query, trigger, triggerTysStr)
 }