+++ sumologic/sumologic_monitors_library_monitor.go
@@ -136,7 +136,7 @@ type MonitorsLibraryMonitor struct {
 	IsDisabled         bool                  `json:"isDisabled"`
 	Status             []string              `json:"status"`
 	GroupNotifications bool                  `json:"groupNotifications"`
-	Playbook           string                `json:"playbook"`
+	Playbook           string                `json:"playbook,omitempty"`
 }
 
 type MonitorQuery struct {