+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2.9.10 (Unreleased)
 
+FEATURES:
+
+* Add a new optional field `playbook` to resource/sumologic_monitor.
+
 ## 2.9.9 (August 12, 2021)
 
 BUG FIXES: